[Refactor] #57 - 토큰 갱신 시 헤더에 Authorization 필요없어서 Inteptor 제거
