+++ data/src/main/java/com/yapp/buddycon/data/di/DataSourceModule.kt
@@ -2,8 +2,6 @@ package com.yapp.buddycon.data.di
 
 import com.yapp.buddycon.data.datasource.remote.login.LoginRemoteDataSource
 import com.yapp.buddycon.data.datasource.remote.login.LoginRemoteDataSourceImpl
-import com.yapp.buddycon.data.datasource.remote.token.RefreshTokenRemoteDataSource
-import com.yapp.buddycon.data.datasource.remote.token.RefreshTokenRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -17,9 +15,4 @@ interface DataSourceModule {
     fun bindsLoginRemoteDataSource(
         loginRemoteDataSourceImpl: LoginRemoteDataSourceImpl
     ): LoginRemoteDataSource
-
-    @Binds
-    fun bindsRefreshTokenRemoteDataSource(
-        refreshTokenRemoteDataSourceImpl: RefreshTokenRemoteDataSourceImpl
-    ): RefreshTokenRemoteDataSource
 }
