+++ data/src/main/java/com/yapp/buddycon/data/di/InterceptorModule.kt
@@ -3,9 +3,7 @@ package com.yapp.buddycon.data.di
 import com.yapp.buddycon.data.BuildConfig
 import com.yapp.buddycon.data.network.BuddyConInterceptorQualifier
 import com.yapp.buddycon.data.network.HttpLoggingInterceptorQualifier
-import com.yapp.buddycon.data.network.RefreshTokenInterceptorQualifier
 import com.yapp.buddycon.data.network.interceptor.BuddyConInterceptor
-import com.yapp.buddycon.data.network.interceptor.RefreshTokenInterceptor
 import com.yapp.buddycon.data.repository.TokenRepositoryImpl
 import com.yapp.buddycon.data.repository.UserRepositoryImpl
 import dagger.Module
@@ -27,14 +25,6 @@ object InterceptorModule {
         HttpLoggingInterceptor().apply {
             level = if (BuildConfig.DEBUG) HttpLoggingInterceptor.Level.BODY else HttpLoggingInterceptor.Level.NONE
         }
-
-    @RefreshTokenInterceptorQualifier
-    @Provides
-    @Singleton
-    fun provideRefreshTokenInterceptor(
-        tokenRepositoryImpl: TokenRepositoryImpl
-    ): Interceptor =
-        RefreshTokenInterceptor(tokenRepositoryImpl)
 
     @BuddyConInterceptorQualifier
     @Provides
