+++ data/src/main/java/com/yapp/buddycon/data/network/api/LoginService.kt
@@ -1,5 +1,6 @@
 package com.yapp.buddycon.data.network.api
 
+import com.yapp.buddycon.data.network.request.RefreshTokenRequest
 import com.yapp.buddycon.data.network.request.UserInfoRequest
 import com.yapp.buddycon.data.network.response.UserInfoResponse
 import retrofit2.http.Body
@@ -10,5 +11,10 @@ interface LoginService {
     @POST("api/v1/login")
     suspend fun requestUserInfo(
         @Body userInfoRequest: UserInfoRequest
+    ): UserInfoResponse
+
+    @POST("api/v1/reissue")
+    suspend fun requestRefreshToken(
+        @Body refreshTokenRequest: RefreshTokenRequest
     ): UserInfoResponse
 }
