--- data/src/main/java/com/yapp/buddycon/data/network/api/RefreshTokenService.kt
@@ -1,13 +1,0 @@
-package com.yapp.buddycon.data.network.api
-
-import com.yapp.buddycon.data.network.request.RefreshTokenRequest
-import com.yapp.buddycon.data.network.response.UserInfoResponse
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface RefreshTokenService {
-    @POST("api/v1/reissue")
-    suspend fun requestRefreshToken(
-        @Body refreshTokenRequest: RefreshTokenRequest
-    ): UserInfoResponse
-}
