+++ data/src/main/java/com/yapp/buddycon/data/network/qualifiers/InterceptorQualifiers.kt
@@ -8,8 +8,4 @@ annotation class HttpLoggingInterceptorQualifier
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
-annotation class RefreshTokenInterceptorQualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
 annotation class BuddyConInterceptorQualifier
