+++ data/src/main/java/com/yapp/buddycon/data/network/qualifiers/OkHttpClientQualifiers.kt
@@ -9,8 +9,4 @@ annotation class LoginClient
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
-annotation class RefreshTokenClient
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
 annotation class BuddyConClient
