+++ data/src/main/java/com/yapp/buddycon/data/network/qualifiers/RetrofitQualifiers.kt
@@ -9,8 +9,4 @@ annotation class LoginRetrofit
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
-annotation class RefreshTokenRetrofit
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
 annotation class BuddyConRetrofit
