+++ data/src/main/java/com/yapp/buddycon/data/repository/UserRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.yapp.buddycon.data.repository
 
 import com.yapp.buddycon.data.datasource.remote.login.LoginRemoteDataSource
-import com.yapp.buddycon.data.datasource.remote.token.RefreshTokenRemoteDataSource
 import com.yapp.buddycon.domain.model.UserInfo
 import com.yapp.buddycon.domain.repository.UserRepository
 import kotlinx.coroutines.flow.*
@@ -9,8 +8,7 @@ import javax.inject.Inject
 
 
 class UserRepositoryImpl @Inject constructor(
-    private val loginRemoteDataSource: LoginRemoteDataSource,
-    private val refreshTokenRemoteDataSource: RefreshTokenRemoteDataSource
+    private val loginRemoteDataSource: LoginRemoteDataSource
 ) : UserRepository {
 
     override fun requestUserInfo(
@@ -24,6 +22,6 @@ class UserRepositoryImpl @Inject constructor(
             .map { it.toModel() }
 
     override fun requestRefreshToken(accessToken: String, refreshToken: String): Flow<UserInfo> =
-        refreshTokenRemoteDataSource.requestRefreshToken(accessToken, refreshToken)
+        loginRemoteDataSource.requestRefreshToken(accessToken, refreshToken)
             .map { it.toModel() }
 }
